Refactor and cleanup in the code
-  Create devConsole/pipeline which contains the various components of the watch/schedule/build/publish pipeline
-  Remove cli/loggers; use console.* for logging messages before the terminal UI starts
-  Emit less messages into appLog
-  Code docs
-  Renaming couple of methods
