+++ tools/applink/src/devConsole/action.ts
@@ -1,14 +1,12 @@
 // (C) 2020 GoodData Corporation
 import { getSourceDescriptor } from "../base/sourceDiscovery";
-import { discoverTargetDependencies } from "../base/targetDiscovery";
+import { getTargetDescriptor } from "../base/targetDiscovery";
 import { TerminalUi } from "./ui/ui";
-import { logInfo, registerLogFn } from "../cli/loggers";
-import { appLogInfo } from "./ui/utils";
 import { GlobalEventBus, sourceInitialized, targetSelected } from "./events";
-import { ChangeDetector } from "./changeDetector";
-import { BuildScheduler } from "./buildScheduler";
-import { PackageBuilder } from "./packageBuilder";
-import { PackagePublisher } from "./publisher";
+import { ChangeDetector } from "./pipeline/changeDetector";
+import { BuildScheduler } from "./pipeline/buildScheduler";
+import { PackageBuilder } from "./pipeline/packageBuilder";
+import { PackagePublisher } from "./pipeline/publisher";
 
 export async function devConsole(targetDir: string): Promise<number> {
     const sourceDescriptor = await getSourceDescriptor(
@@ -19,18 +17,13 @@ export async function devConsole(targetDir: string): Promise<number> {
         return 1;
     }
 
-    const targetDescriptor = discoverTargetDependencies(targetDir, sourceDescriptor);
+    const targetDescriptor = getTargetDescriptor(targetDir, sourceDescriptor);
 
     if (!targetDescriptor.dependencies.length) {
-        logInfo(`The target project does not have any dependencies on the SDK. There is nothing to do.`);
+        console.info("The target project does not have any dependencies on the SDK. There is nothing to do.");
 
         return 1;
     }
-
-    /*
-     * Register log function so that all messages land in the application's log
-     */
-    registerLogFn(appLogInfo);
 
     /*
      * Initialize the terminal UI - this will make the app run forever until user triggers exit
